+++ src/videobandwidthestimator.h
@@ -91,6 +91,6 @@ unsigned int ortp_video_bandwidth_estimator_get_trust(OrtpVideoBandwidthEstimato
 
 void ortp_video_bandwidth_estimator_process_packet(OrtpVideoBandwidthEstimator *vbe, uint32_t sent_timestamp, const struct timeval *recv_timestamp, int msgsize, bool_t is_last);
 
-int ortp_video_bandwidth_estimator_get_estimated_available_bandwidth(OrtpVideoBandwidthEstimator *vbe);
+float ortp_video_bandwidth_estimator_get_estimated_available_bandwidth(OrtpVideoBandwidthEstimator *vbe);
 
 #endif
